requirements.txt
PyPDF2>=3.0.0
pytesseract>=0.3.10
Pillow>=9.0.0
openai>=1.0.0
fpdf>=1.7.2
streamlit>=1.20.0
httpx>=0.23.0
pandas>=1.5.0
plotly>=5.10.0
glob2>=0.7